pyproject.toml
[tool.poetry]
name = "brainhack_western"
version = "0.1.0"
description = "Utility for determining overlapping scheduled events"
authors = ["Peter Van Dyken"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
networkx = "^2.6.3"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
